remove neuapp
